홈에서 검색 기능- 빌드 오류 수정
