+++ src/app/category-settings/page.tsx
@@ -17,14 +17,7 @@ import {
   verticalListSortingStrategy,
 } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
-import {
-  Box,
-  List,
-  ListItem,
-  IconButton,
-  Popover,
-  Button,
-} from '@mui/material';
+import { Box, List, ListItem, IconButton, Popover } from '@mui/material';
 import Layout from '../../Components/LayoutDir/Layout';
 import DragHandleIcon from '@mui/icons-material/DragHandle';
 import NotificationsIcon from '@mui/icons-material/Notifications';
